+++ DG_backend/app/schemas/order_schema.py
@@ -5,11 +5,11 @@ class OrderItemSchema(BaseModel):
     menu_item_id: str
     name: str
     quantity: int
-    price: int
+    price: float
     instructions: optional[str] = None
 
 class OrderCreate(BaseModel):
     items: List[OrderItemSchema]
-    total_amount: int
+    total_amount: float
     delivery_address: str
     payment_method: str
